pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "robosafe-sentinel"
version = "0.1.0"
description = "Système de sécurité intelligent pour cellules robotisées industrielles"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.10"
authors = [
    {name = "Preventera / GenAISafety", email = "dev@genaisafety.com"}
]
keywords = ["robotics", "safety", "ai", "industrial", "hse", "cobot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Manufacturing",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: System :: Hardware",
]

dependencies = [
    # Core
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
    
    # Async & Networking
    "asyncio>=3.4.3",
    "aiohttp>=3.9.0",
    "httpx>=0.25.0",
    
    # Industrial Protocols
    "pycomm3>=1.2.0",           # EtherNet/IP (Fanuc, Rockwell)
    "python-snap7>=1.3",         # Siemens S7
    "pymodbus>=3.6.0",           # Modbus TCP
    "opcua>=0.98.0",             # OPC-UA
    
    # Computer Vision
    "opencv-python>=4.8.0",
    "numpy>=1.26.0",
    "ultralytics>=8.0.0",        # YOLO for detection
    
    # API & Web
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    
    # Monitoring & Logging
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "rich>=13.7.0",
    
    # Data & Storage
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "redis>=5.0.0",
    "influxdb-client>=1.38.0",
    
    # Utilities
    "click>=8.1.0",
    "schedule>=1.2.0",
    "watchdog>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "hypothesis>=6.92.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    "ipython>=8.18.0",
]

docs = [
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]

[project.scripts]
robosafe = "robosafe.cli:main"
robosafe-server = "robosafe.api.server:run"

[project.urls]
Homepage = "https://github.com/Preventera/RoboSafeSentinel"
Documentation = "https://docs.genaisafety.com/robosafe"
Repository = "https://github.com/Preventera/RoboSafeSentinel.git"
Issues = "https://github.com/Preventera/RoboSafeSentinel/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
robosafe = ["py.typed"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
minversion = "7.0"
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow",
    "integration: marks tests requiring real hardware",
    "unit: marks unit tests",
]

[tool.coverage.run]
source = ["src/robosafe"]
branch = true
omit = ["*/tests/*", "*/__init__.py"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
